WIP: Fix travis
If you don't build boost with c++14, it fails to link because of missing
symbols of system_error_category and generic_error_category. I tried
fixing the current approach of using External_Project in CMake, but I
couldn't figure out, why it didn't accept my flags or where in the
toolchain it went wrong. Boost automatically switches to C++14 with gcc6
and up, but not on older releases. But passing C++14 explicitly didn't
seem to work with any incantation I tried.

My solution now is to replace the External_Project machinery with cget.
Cget requirement files are a lot easier to read in my opinion and so I
think, they will be easier to maintain. It also seems to work, when you
initialize the toolchain with C++14 and fixes the boost issues. I hope
it didn't break anything.