Add man page
This adds a man page which lists command line options, key bindings and
commands. A new CMake option `MAN` is added, enabled by default on
`UNIX` but not `APPLE`.

The man pages are generated with either `a2x` (from asciidoc) or
`asciidoctor`. If both are not found, a fatal error is signaled.