Show only fonts that can render the grin emoji in the emoji font list
Users often are confused what fonts to pick. However this approach is
way too slow. We need to somehow prefilter the list, but for that we
need to know what writing system emojis belong in or some other quick
filter, that we can apply before the QRawFont and possibly the QFont.